# j-lawyer-server/j-lawyer-io/src/main/webapp/swagger-ui/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "8",
    "title": "j-lawyer-io"
  },
  "basePath": "/j-lawyer-io/rest",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/v8/timesheets/cases/{caseId}/positions/last": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getLastPositionsForCase",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RestfulTimesheetPositionV8"
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/open": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getOpenTimesheets",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RestfulTimesheetV8"
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/positions/running": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getRunningPositions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RestfulTimesheetPositionV8"
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/positions/running/count": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getRunningPositionsCount",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getTimesheet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetV8"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}/positions": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getTimesheetPositions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RestfulTimesheetPositionV8"
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      },
      "post": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "addPosition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}/positions/start": {
      "post": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "startPosition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}/positions/{positionId}": {
      "put": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "updatePosition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "deletePosition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}/positions/{positionId}/stop": {
      "put": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "stopPosition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RestfulTimesheetPositionV8"
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "404": {
            "description": "Not Found",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    },
    "/v8/timesheets/{timesheetId}/templates": {
      "get": {
        "tags": [
          "Timesheets"
        ],
        "summary": "",
        "description": "",
        "operationId": "getTimesheetTemplates",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=utf-8"
        ],
        "parameters": [
          {
            "name": "timesheetId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RestfulTimesheetPositionTemplateV8"
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "headers": {}
          },
          "403": {
            "description": "User not authenticated",
            "headers": {}
          },
          "500": {
            "description": "Internal Server Error",
            "headers": {}
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic",
      "description": "HTTP Basic Authentication. Works over HTTP and HTTPS"
    }
  },
  "definitions": {
...
    "RestfulTimesheetPositionTemplateV8": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "unitPrice": {
          "type": "number"
        },
        "taxRate": {
          "type": "number"
        }
      }
    },
    "RestfulTimesheetPositionV8": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "principal": {
          "type": "string"
        },
        "started": {
          "type": "string"
        },
        "stopped": {
          "type": "string"
        },
        "unitPrice": {
          "type": "number"
        },
        "taxRate": {
          "type": "number"
        },
        "total": {
          "type": "number"
        },
        "timesheetId": {
          "type": "string"
        },
        "invoiceId": {
          "type": "string"
        },
        "running": {
          "type": "boolean"
        }
      }
    },
    "RestfulTimesheetV8": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "interval": {
          "type": "integer",
          "format": "int32"
        },
        "limited": {
          "type": "boolean"
        },
        "limit": {
          "type": "number"
        },
        "percentageDone": {
          "type": "number",
          "format": "float"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "caseId": {
          "type": "string"
        },
        "caseFileNumber": {
          "type": "string"
        },
        "caseName": {
          "type": "string"
        }
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}